Refactor - CoreDataManager
